Issue #2326: Add category blacklist ids & SharedPreferences
