+++ app/src/test/java/org/mozilla/tv/firefox/utils/SettingsTest.kt
@@ -8,7 +8,8 @@ import io.mockk.MockKAnnotations
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.mockkStatic
-import org.junit.Assert.*
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
 import org.mozilla.tv.firefox.components.locale.LocaleManager
